+++ src/pages/ItemsPage/components/SortDropdown/SortDropdown.module.scss
@@ -1,0 +1,71 @@
+@use "../../../../styles/variables" as var;
+@use "../../../../styles/mixin" as mixin;
+
+.sortSelectBox {
+  $height: 42px;
+
+  &__current {
+    display: block;
+    width: 100%;
+    padding: 0 42px 0 20px;
+    height: $height;
+    border: 1px solid var.$gray200;
+    font-size: 16px;
+    text-align: left;
+    color: var.$gray800;
+    background: url("../../../../assets/images/icons/ic_select_arrow.svg") right
+      20px center no-repeat;
+    border-radius: 12px;
+
+    @include mixin.mobile {
+      width: $height;
+      padding: 0;
+      background: url("../../../../assets/images/icons/ic_sort.svg") center
+        center no-repeat;
+
+      span {
+        display: none;
+      }
+    }
+  }
+
+  &__list {
+    position: absolute;
+    top: 100%;
+    right: 0;
+    width: 130px;
+    margin-top: 8px;
+    z-index: 1;
+    overflow: hidden;
+
+    ul {
+      border: 1px solid var.$gray200;
+      background: var.$white;
+      border-radius: 12px;
+      transform: translateY(-100%);
+      animation: showSortList 0.3s ease forwards;
+
+      @keyframes showSortList {
+        100% {
+          transform: translateY(0);
+        }
+      }
+
+      li {
+        border-bottom: 1px solid var.$gray200;
+
+        &:last-of-type {
+          border-bottom: none;
+        }
+
+        button {
+          display: block;
+          width: 100%;
+          height: $height;
+          font-size: 16px;
+          color: var.$gray800;
+        }
+      }
+    }
+  }
+}
